Business Process Automation
The automation I made connects to the database via MSSQL. The basic logic of the application is that it allows the business to do everything from small SMEs to large factories, from profit and loss to the number of personnel and the shift of the employees. While developing my project, I used C#, MSSQL and devexpress as a framework.